Organize and centralize HandlerExtensions (#3716)
