+++ src/Core/src/Platform/ElementExtensions.cs
@@ -1,0 +1,120 @@
+﻿using System;
+#if __IOS__ || MACCATALYST
+using NativeView = UIKit.UIView;
+using BasePlatformType = Foundation.NSObject;
+using PlatformWindow = UIKit.UIWindow;
+using PlatformApplication = UIKit.UIApplicationDelegate;
+#elif MONOANDROID
+using NativeView = Android.Views.View;
+using BasePlatformType = Android.Content.Context;
+using PlatformWindow = Android.App.Activity;
+using PlatformApplication = Android.App.Application;
+#elif WINDOWS
+using NativeView = Microsoft.UI.Xaml.FrameworkElement;
+using BasePlatformType = WinRT.IWinRTObject;
+using PlatformWindow = Microsoft.UI.Xaml.Window;
+using PlatformApplication = Microsoft.UI.Xaml.Application;
+#elif NETSTANDARD || (NET6_0 && !IOS && !ANDROID)
+using NativeView = System.Object;
+using BasePlatformType = System.Object;
+using INativeViewHandler = Microsoft.Maui.IViewHandler;
+using PlatformWindow = System.Object;
+using PlatformApplication = System.Object;
+#endif
+
+namespace Microsoft.Maui.Platform
+{
+	public static partial class ElementExtensions
+	{
+		public static IElementHandler ToHandler(this IElement view, IMauiContext context)
+		{
+			_ = view ?? throw new ArgumentNullException(nameof(view));
+			_ = context ?? throw new ArgumentNullException(nameof(context));
+
+			//This is how MVU works. It collapses views down
+			if (view is IReplaceableView ir)
+				view = ir.ReplacedView;
+
+			var handler = view.Handler;
+
+			if (handler?.MauiContext != null && handler.MauiContext != context)
+				handler = null;
+
+			if (handler == null)
+				handler = context.Handlers.GetHandler(view.GetType());
+
+			if (handler == null)
+				throw new Exception($"Handler not found for view {view}.");
+
+			handler.SetMauiContext(context);
+
+			view.Handler = handler;
+
+			if (handler.VirtualView != view)
+				handler.SetVirtualView(view);
+
+			return handler;
+		}
+
+		internal static NativeView? GetNative(this IElement view, bool returnWrappedIfPresent)
+		{
+			if (view.Handler is INativeViewHandler nativeHandler && nativeHandler.NativeView != null)
+				return nativeHandler.NativeView;
+
+			return view.Handler?.NativeView as NativeView;
+
+		}
+
+		internal static NativeView ToNative(this IElement view, IMauiContext context, bool returnWrappedIfPresent)
+		{
+			var nativeView = view.ToNative(context);
+
+			if (view.Handler is INativeViewHandler nativeHandler && nativeHandler.NativeView != null)
+				return nativeHandler.NativeView;
+
+			return nativeView;
+
+		}
+
+		public static NativeView ToNative(this IElement view, IMauiContext context)
+		{
+			var handler = view.ToHandler(context);
+
+			if (handler.NativeView is not NativeView result)
+			{
+				throw new InvalidOperationException($"Unable to convert {view} to {typeof(NativeView)}");
+			}
+			return result;
+		}
+
+		static void SetHandler(this BasePlatformType nativeElement, IElement element, IMauiContext context)
+		{
+			_ = nativeElement ?? throw new ArgumentNullException(nameof(nativeElement));
+			_ = element ?? throw new ArgumentNullException(nameof(element));
+			_ = context ?? throw new ArgumentNullException(nameof(context));
+
+			var handler = element.Handler;
+			if (handler?.MauiContext != null && handler.MauiContext != context)
+				handler = null;
+
+			if (handler == null)
+				handler = context.Handlers.GetHandler(element.GetType());
+
+			if (handler == null)
+				throw new Exception($"Handler not found for window {element}.");
+
+			handler.SetMauiContext(context);
+
+			element.Handler = handler;
+
+			if (handler.VirtualView != element)
+				handler.SetVirtualView(element);
+		}
+
+		public static void SetApplicationHandler(this PlatformApplication nativeApplication, IApplication application, IMauiContext context) =>
+			SetHandler(nativeApplication, application, context);
+
+		public static void SetWindowHandler(this PlatformWindow nativeWindow, IWindow window, IMauiContext context) =>
+			SetHandler(nativeWindow, window, context);
+	}
+}
