+++ src/Core/tests/DeviceTests/Handlers/Element/ElementTests.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Microsoft.Maui.DeviceTests.Stubs;
+using Xunit;
+
+namespace Microsoft.Maui.DeviceTests
+{
+	public partial class ElementTests : HandlerTestBase
+	{
+		[Fact]
+		public void ElementToHandlerReturnsIElementHandler()
+		{
+			var handler = new ElementStub().ToHandler(MauiContext);
+			Assert.NotNull(handler);
+		}
+	}
+}
