+++ src/Core/tests/DeviceTests/Handlers/HandlerTestBaseOfT.Windows.cs
