+++ src/Core/tests/DeviceTests/Stubs/ElementHandlerStub.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using Microsoft.Maui.DeviceTests.Stubs;
+
+namespace Microsoft.Maui.DeviceTests
+{
+	public class ElementHandlerStub : ElementHandler<ElementStub, object>
+	{
+		public ElementHandlerStub() : base(ElementHandler.ElementMapper)
+		{
+
+		}
+
+		protected override object CreateNativeElement() => new Object();
+	}
+}
