+++ src/Core/tests/DeviceTests/Stubs/ElementStub.cs
@@ -1,0 +1,9 @@
+﻿namespace Microsoft.Maui.DeviceTests.Stubs
+{
+	public class ElementStub : IElement
+	{
+		public IElement Parent { get; set; }
+
+		public IElementHandler Handler { get; set; }
+	}
+}
